Move tests module to reflect package restructure
